better memoize support
If no mmz-package is loaded, the libraries act as before, they simply write the infos to the aux file.
If any mmz-package is loaded, it writes the into to the aux file if mmz is aborted or if mmz is disabled. If mmz is active and the picture is stable, the info will not written to the aux file because the info isn't needed next compilation run.